+++ api_tests/replication_api_test.py
@@ -37,7 +37,7 @@ def test__create_replication_ok(base_url, session, bucket_name, replication_name
         "settings": {
             "dst_bucket": "dst_bucket",
             "dst_host": "http://localhost:9000",
-            "dst_token": "",
+            "dst_token": "***",
             "entries": [],
             "exclude": {},
             "include": {},
@@ -103,7 +103,7 @@ def test__update_replication_ok(base_url, session, bucket_name, replication_name
         "settings": {
             "dst_bucket": bucket_name,
             "dst_host": "http://localhost:9001",
-            "dst_token": "",
+            "dst_token": "***",
             "entries": [],
             "exclude": {},
             "include": {},